+++ src/assets.test.ts
